requirements.txt
mcp
sqlalchemy
pandas
python-dotenv
anthropic
openpyxl
xlsxwriter